+++ src/main/java/com/estacionamento/DTO/VeiculoSaidaDto.java
@@ -4,5 +4,5 @@ import com.estacionamento.domain.VeiculoTipo;
 
 import java.time.LocalDateTime;
 
-public record VeiculoSaidaDto (VeiculoTipo tipoVeiculo, String placa, String modelo, LocalDateTime entrada, LocalDateTime saida, Long periodoEmMinutos, String valor) {
+public record VeiculoSaidaDto (VeiculoTipo tipoVeiculo, String placa, String modelo, LocalDateTime entrada, LocalDateTime saida, Integer periodoEmMinutos, String valor) {
 }
